BC-4787 - move logic to mapper
