+++ apps/server/src/modules/school/repo/mikro-orm/mapper/school.entity.mapper.ts
@@ -1,3 +1,6 @@
+import { EntityData } from '@mikro-orm/core';
+import { EntityManager } from '@mikro-orm/mongodb';
+import { FederalStateEntity, SchoolYearEntity } from '@shared/domain/entity';
 import { SchoolEntity } from '@shared/domain/entity/school.entity';
 import { SchoolFactory } from '@src/modules/school/domain/factory';
 import { School } from '../../../domain';
@@ -45,5 +48,36 @@ export class SchoolEntityMapper {
 		const schools = schoolEntities.map((entity) => SchoolEntityMapper.mapToDo(entity));
 
 		return schools;
+	}
+
+	public static mapToEntityProperties(domainObject: School, em: EntityManager): EntityData<SchoolEntity> {
+		const props = domainObject.getProps();
+		const federalState = props.federalState ? em.getReference(FederalStateEntity, props.federalState?.id) : undefined;
+		const features = Array.from(props.features);
+		const currentYear = props.currentYear ? em.getReference(SchoolYearEntity, props.currentYear?.id) : undefined;
+		const county = props.county ? CountyEmbeddableMapper.mapToEntity(props.county) : undefined;
+
+		const schoolEntityProps = {
+			name: props.name,
+			officialSchoolNumber: props.officialSchoolNumber,
+			externalId: props.externalId,
+			previousExternalId: props.previousExternalId,
+			inMaintenanceSince: props.inMaintenanceSince,
+			inUserMigration: props.inUserMigration,
+			purpose: props.purpose,
+			logo_dataUrl: props.logo_dataUrl,
+			logo_name: props.logo_name,
+			fileStorageType: props.fileStorageType,
+			language: props.language,
+			timezone: props.timezone,
+			permissions: props.permissions,
+			enableStudentTeamCreation: props.enableStudentTeamCreation,
+			federalState,
+			features,
+			currentYear,
+			county,
+		};
+
+		return schoolEntityProps;
 	}
 }
